fixed ids being int instead of long

+++ models/src/main/java/ar/edu/itba/paw/models/Rating.java
@@ -2,12 +2,12 @@ package ar.edu.itba.paw.models;
 
 public class Rating {
     private Double rating;
-    private int doctorId;
-    private int patientId;
-    private int appointmentId;
+    private long doctorId;
+    private long patientId;
+    private long appointmentId;
     private String comment;
     private long id;
-    public Rating(Double rating, int doctorId, int patientId, int appointmentId, String comment, long id) {
+    public Rating(Double rating, long doctorId, long patientId, long appointmentId, String comment, long id) {
         this.rating = rating;
         this.doctorId = doctorId;
         this.patientId = patientId;
@@ -20,15 +20,15 @@ public class Rating {
         this.rating = rating;
     }
 
-    public void setDoctorId(int doctorId) {
+    public void setDoctorId(long doctorId) {
         this.doctorId = doctorId;
     }
 
-    public void setPatientId(int patientId) {
+    public void setPatientId(long patientId) {
         this.patientId = patientId;
     }
 
-    public void setAppointmentId(int appointmentId) {
+    public void setAppointmentId(long appointmentId) {
         this.appointmentId = appointmentId;
     }
 
@@ -44,15 +44,15 @@ public class Rating {
         return rating;
     }
 
-    public int getDoctorId() {
+    public long getDoctorId() {
         return doctorId;
     }
 
-    public int getPatientId() {
+    public long getPatientId() {
         return patientId;
     }
 
-    public int getAppointmentId() {
+    public long getAppointmentId() {
         return appointmentId;
     }
